First version of adding internationalize to the bundle
