+++ src/modules/cookie.js
@@ -1,5 +1,6 @@
 import { CookieConsent } from '@sandstreamdev/cookieconsent';
 import '@sandstreamdev/cookieconsent/build/cookieconsent.min.css';
+import {i18n_map, preferredLang} from './internationalize'
 export default function() {
     new CookieConsent({
         "palette": {
